cdp: Rename module interface file #82
